+++ test/testbed.hs
@@ -1,9 +1,9 @@
 module TestBed where
 
 import Data.IntMap.Bound
-import Data.IntMap.Bound.Base
 import Data.IntMap.Bound.Verifier
 import Data.IntMap.Hack.Analysis
+import Data.IntMap.Hack.Base
 import Data.IntMap.Hack.Deprecated
 import Data.IntMap.Hack.Tool
 import Debug.Shortcut